sepparate the app initialization
